// src/shared/types/event.ts
export enum EventTimelineStatus {
  NOT_STARTED = 'not_started',
  STARTED = 'started',
  FINISHED = 'finished',
}

export enum EventCardBadgeStatus {
  EXHIBIT_CLOSED = 'EXHIBIT_CLOSED',
  EXHIBIT_NOT_STARTED = 'EXHIBIT_NOT_STARTED',
  AUCTION_LIVE = 'AUCTION_LIVE',
  TICKET_ON_SALE = 'TICKET_ON_SALE',
  NONE = 'NONE',
}